pyproject.toml
[project]
name = "debridbridge"
version = "0.1.0"
description = "Single-container Real-Debrid bridge replacing Zurg + Rclone + Decypharr"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "guessit>=3.8.0",
    "parse-torrent-title>=0.6.0",
    "httpx>=0.27.0",
    "wsgidav>=4.3.0",
    "pydantic-settings>=2.5.0",
    "apscheduler>=3.10.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.9",
    "cheroot>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
